Extract exchange rate widget to a separate component
